refactor emal helpers into provider pattern
