+++ api/modules/billing/billing.dependencies.ts
@@ -5,6 +5,7 @@ import { BillingDataProvider } from './billing.dataprovider';
 import { StripeProvider } from '../../providers/stripe';
 import { useLogger } from '../../lib/loggingMiddleware';
 import { NextFunction, Request, Response } from 'express';
+import { EmailProvider } from '@/providers/emailProvider';
 
 const container = createContainer({
   injectionMode: InjectionMode.CLASSIC,
@@ -15,6 +16,7 @@ container.register({
   billingService: asClass(BillingService).scoped(),
   billingDataProvider: asClass(BillingDataProvider).scoped(),
   stripeProvider: asClass(StripeProvider).singleton(),
+  emailProvider: asClass(EmailProvider).singleton(),
   logger: asFunction(useLogger).scoped(),
 });
